Start creating the frontend

+++ backend/controllers/product.controller.js
@@ -56,6 +56,13 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  // if the user passed an invalid id
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res
@@ -63,6 +70,6 @@ export const deleteProduct = async (req, res) => {
       .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
     console.log("Error deleting product:", error.message);
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
